Data.X509.CertificateStore

Documentation

data CertificateStore

A Collection of certificate or store of certificates.

makeCertificateStore :: [SignedCertificate] -> CertificateStore

Create a certificate store out of a list of X509 certificate

readCertificateStore :: FilePath -> IO (Maybe CertificateStore)

Create certificate store by reading certificates from file or directory

This function can be used to read multiple certificates from either single file (multiple PEM formatted certificates concanated) or directory (one certificate per file, file names are hashes from certificate).

readCertificates :: FilePath -> IO [SignedCertificate]

Queries

findCertificate :: DistinguishedName -> CertificateStore -> Maybe SignedCertificate

Find a certificate using the subject distinguished name

listCertificates :: CertificateStore -> [SignedCertificate]

List all certificates in a store